feat(auth): Refactored authorization service and fixed parameter issues in controllers
- Restructured Redis-related files:
  - Moved  to  directory
  - Moved  and  to  directory
- Fixed incorrect parameter passing in services
- Updated  and  to correctly handle request parameters
- Modified  and  for improved data handling
- Applied fixes in  and  to align with corrected logic

These changes improve authorization handling, ensure services receive the correct parameters, and align the Redis cache structure for better maintainability.
