+++ user_managment_core/src/user/contracts/user.repository.Interface.ts
