+++ user_managment_core/src/user/user.controllers/auth.controller.ts
@@ -4,7 +4,7 @@ import { ApiTags, ApiOperation } from '@nestjs/swagger';
 import { CreateAuthDto } from '../dto/create-auth.dto';
 import { JwtAuthGuard } from '../libs/jwt-auth.guard';
 import { GetUser } from '../libs/decorators/getUser';
-import { IuserReadService } from '../libs/contracts/user.repository.Interface';
+import { IuserReadService } from '../contracts/user.repository.Interface';
 import { Body, Controller, Get, Inject, Injectable, Param, Post, Scope, UnauthorizedException, UseGuards } from '@nestjs/common';
 
 @Injectable({ scope: Scope.DEFAULT }) 
@@ -31,8 +31,8 @@ export class AuthController {
     if (!user) {
       throw new UnauthorizedException('User not found');
     }
-    console.log('getProfileByemail',body)
-    return await this.UseReadService.getProfileByEmailService(user.email);
+  
+    return await this.UseReadService.getProfileByEmailService(body.email);
 
   }
